test: add init tests for cli
